refactor(client): extract connection establishment into client/connect.rs
Move the entire impl Client<Disconnected> block (1,009 lines) into a
dedicated submodule. This includes connect, try_connect, connect_tds_8,
connect_tds_7x, connect_no_tls, build_prelogin, build_login7,
send_prelogin, receive_prelogin, send_login7, process_login_response,
and process_env_change.

client/mod.rs: 2,941 → 1,930 lines (-34%)
New client/connect.rs: 1,039 lines

Zero public API changes. Zero warnings across all feature configurations.
